+++ lib/extensionConfiguration.js
@@ -4,7 +4,7 @@ const { generateTrackerAllowlistRules } = require('./trackerAllowlist')
 
 /**
  * @typedef {object} generateExtensionConfigurationRulesetResult
- * @property {object[]} ruleset
+ * @property {import('./utils.js').DNRRule[]} ruleset
  *   The generated declarativeNetRequest ruleset.
  * @property {object} matchDetailsByRuleId
  *   Rule ID -> match details.
