+++ lib/trackerAllowlist.js
@@ -1,9 +1,11 @@
 /** @module trackerAllowlist */
 
-const { storeInLookup } = require('./utils')
 const {
-    getTrackerEntryDomain, generateDNRRule
-} = require('./tds')
+    processPlaintextTrackerRule,
+    storeInLookup,
+    generateDNRRule,
+    getTrackerEntryDomain
+} = require('./utils')
 
 // Priority that the Tracker Blocking Allowlist declarativeNetRequest rules
 // start from.
@@ -24,7 +26,7 @@ const MAXIMUM_RULES_PER_TRACKER_ENTRY = (CEILING_PRIORITY - BASELINE_PRIORITY) /
  * details for the given trackerAllowlist configuration.
  * @param {object} extensionConfiguration
  *   The extension configuration.
- * @return {Generator<[import('./tds').DNRRule, object]>}
+ * @return {Generator<[import('./utils').DNRRule, object]>}
  */
 function* generateTrackerAllowlistRules ({ features: { trackerAllowlist } }) {
     // No allowlisted trackers.
@@ -83,7 +85,7 @@ function* generateTrackerAllowlistRules ({ features: { trackerAllowlist } }) {
         let priority = BASELINE_PRIORITY
         for (let i = trackerEntryRules.length - 1; i >= 0; i--) {
             let {
-                rule: urlFilter,
+                rule: trackerRule,
                 domains: initiatorDomains,
                 reason: allowlistReason
             } = trackerEntryRules[i]
@@ -95,19 +97,16 @@ function* generateTrackerAllowlistRules ({ features: { trackerAllowlist } }) {
                 initiatorDomains = null
             }
 
-            // Tracker Blocking Allowlist entries do not support regular
-            // expression matching, so processTrackerRule cannot be used here.
-            if (urlFilter.startsWith(trackerDomain)) {
-                urlFilter = '||' + urlFilter
-            }
+            const {
+                urlFilter,
+                matchCase
+            } = processPlaintextTrackerRule(trackerDomain, trackerRule)
 
             const rule = generateDNRRule({
                 priority,
                 actionType: 'allow',
                 urlFilter,
-                // Note: In the future it would be nice to avoid
-                //       case-insensitive matching unless necessary.
-                matchCase: false,
+                matchCase,
                 requestDomains,
                 excludedRequestDomains,
                 initiatorDomains
